changed unsigned int to size_t due to registry problems

+++ ugbase/common/util/string_util.cpp
@@ -221,27 +221,27 @@ const unsigned int cost_sub = 1;
  * (check copyright or recreate!)
  *
  */
-unsigned int LevenshteinDistance( const std::string& s1, const std::string& s2 )
+size_t LevenshteinDistance( const std::string& s1, const std::string& s2 )
 {
-  unsigned int n1 = s1.length();
-  unsigned int n2 = s2.length();
+  size_t n1 = s1.length();
+  size_t n2 = s2.length();
 
-  unsigned int* p = new unsigned int[ n2+1 ];
-  unsigned int* q = new unsigned int[ n2+1 ];
-  unsigned int* r;
+  size_t* p = new size_t[ n2+1 ];
+  size_t* q = new size_t[ n2+1 ];
+  size_t* r;
 
   p[ 0 ] = 0;
-  for( unsigned int j = 1; j <= n2; ++j )
+  for( size_t j = 1; j <= n2; ++j )
     p[ j ] = p[ j-1 ] + cost_ins;
 
-  for( unsigned int i = 1; i <= n1; ++i )
+  for( size_t i = 1; i <= n1; ++i )
     {
       q[ 0 ] = p[ 0 ] + cost_del;
-      for( unsigned int j = 1; j <= n2; ++j )
+      for( size_t j = 1; j <= n2; ++j )
         {
-          unsigned int d_del = p[ j   ] + cost_del;
-          unsigned int d_ins = q[ j-1 ] + cost_ins;
-          unsigned int d_sub = p[ j-1 ] + ( s1[i-1] == s2[j-1] ? 0 : cost_sub );
+          size_t d_del = p[ j   ] + cost_del;
+          size_t d_ins = q[ j-1 ] + cost_ins;
+          size_t d_sub = p[ j-1 ] + ( s1[i-1] == s2[j-1] ? 0 : cost_sub );
           q[ j ] = std::min( std::min( d_del, d_ins ), d_sub );
       }
       r = p;
@@ -249,7 +249,7 @@ unsigned int LevenshteinDistance( const std::string& s1, const std::string& s2 )
       q = r;
     }
 
-  unsigned int tmp = p[ n2 ];
+  size_t tmp = p[ n2 ];
   delete[] p;
   delete[] q;
